Port Linux/macOS to 64-bit, using new CMake build system

+++ idlib/Heap.h
@@ -214,7 +214,9 @@ type *idBlockAlloc<type,blockSize>::Alloc( void ) {
 
 template<class type, int blockSize>
 void idBlockAlloc<type,blockSize>::Free( type *t ) {
-	element_t *element = (element_t *)( ( (unsigned char *) t ) - ( (int) &((element_t *)0)->t ) );
+	// flibit: 64 bit fix, changed int to intptr_t
+	element_t *element = (element_t *)( ( (unsigned char *) t ) - ( (intptr_t) &((element_t *)0)->t ) );
+	// flibit end
 	element->next = free;
 	free = element;
 	active--;
